Task 2: Implement fetchProductsThen using .then() and .catch()
Created a function that uses .then() and .catch() to pull the product information from the link in order to display the names to the console and raise an error if there is a problem fetching a product. The function also stores the product to the dashboard as a list that contains the name and price and prints it to the page as a list.